pyproject.toml
[project]
name = "event-finder-api"
version = "0.1.0"
description = "Event Finder API built with FastAPI"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Nik Gospodinov"},
]
dependencies = [
    "fastapi==0.116.1",
    "uvicorn==0.35.0",
    "pydantic-settings==2.10.1",
    "langchain-google-genai==2.1.8",
    "langchain-ollama==0.3.4"
]

[project.optional-dependencies]
dev = [
    "mypy==1.17.0",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "black==23.11.0",
    "isort==5.12.0",
    "flake8==6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
select = ["E", "F", "W", "C90"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "venv",
    ".mypy_cache",
    ".pytest_cache",
]
